refactor(trade): extract dialog components to sibling files
fix #418
