+++ src/features/trade/components/SettlementTradeRoutesPanel/CancelConfirmDialog.test.tsx
@@ -1,0 +1,202 @@
+import { QueryClient, QueryClientProvider } from "@tanstack/react-query";
+import { render, screen, waitFor } from "@testing-library/react";
+import userEvent from "@testing-library/user-event";
+import { beforeEach, describe, expect, it, vi } from "vitest";
+
+import { CancelConfirmDialog } from "./CancelConfirmDialog";
+
+import type { TradeRoute } from "../../types/tradeRouteTypes";
+
+const { requireSupabaseClient } = vi.hoisted(() => ({
+  requireSupabaseClient: vi.fn<() => unknown>(),
+}));
+
+vi.mock("@/lib/supabase", () => ({
+  requireSupabaseClient,
+}));
+
+const { toastError, toastSuccess } = vi.hoisted(() => ({
+  toastError: vi.fn<(message: string) => void>(),
+  toastSuccess:
+    vi.fn<(message: string, options?: { description?: string }) => void>(),
+}));
+
+vi.mock("sonner", () => ({
+  toast: {
+    error: toastError,
+    success: toastSuccess,
+  },
+}));
+
+const ROUTE_ID = "00000000-0000-0000-0000-000000000010";
+const SETTLEMENT_ID = "00000000-0000-0000-0000-000000000001";
+const DEST_SETTLEMENT_ID = "00000000-0000-0000-0000-000000000002";
+const CITIZEN_ID = "00000000-0000-0000-0000-000000000040";
+
+function makeRoute(): TradeRoute {
+  return {
+    id: ROUTE_ID,
+    originSettlementId: SETTLEMENT_ID,
+    originSettlementName: "Home Settlement",
+    originNationName: "Home Nation",
+    destinationSettlementId: DEST_SETTLEMENT_ID,
+    destinationSettlementName: "Far Settlement",
+    destinationNationName: "Far Nation",
+    resourceId: "00000000-0000-0000-0000-000000000030",
+    resourceName: "Grain",
+    quantityPerTransition: 10,
+    status: "active",
+    originApprovalStatus: "approved",
+    destinationApprovalStatus: "approved",
+    replacementForTradeRouteId: null,
+    pauseReasonLastTransition: null,
+    proposedByCitizenId: CITIZEN_ID,
+    originApprovedByCitizenId: CITIZEN_ID,
+    destinationApprovedByCitizenId: CITIZEN_ID,
+    createdAt: "2026-06-01T00:00:00.000Z",
+    updatedAt: "2026-06-01T00:00:00.000Z",
+  };
+}
+
+function makeRpcMock(): ReturnType<typeof vi.fn> {
+  return vi.fn((fn: string) => {
+    if (fn === "cancel_trade_route") {
+      return {
+        maybeSingle: vi.fn().mockResolvedValue({
+          data: {
+            id: ROUTE_ID,
+            origin_settlement_id: SETTLEMENT_ID,
+            destination_settlement_id: DEST_SETTLEMENT_ID,
+            status: "cancelled",
+          },
+          error: null,
+        }),
+      };
+    }
+    throw new Error(`Unexpected RPC: ${fn}`);
+  });
+}
+
+function renderDialog({
+  onClose = vi.fn<() => void>(),
+  traderCount = 0,
+  rpcMock = makeRpcMock(),
+}: {
+  readonly onClose?: () => void;
+  readonly traderCount?: number;
+  readonly rpcMock?: ReturnType<typeof vi.fn>;
+} = {}): {
+  readonly onClose: () => void;
+  readonly rpcMock: ReturnType<typeof vi.fn>;
+} {
+  requireSupabaseClient.mockReturnValue({ rpc: rpcMock });
+  const queryClient = new QueryClient({
+    defaultOptions: { queries: { retry: false } },
+  });
+  render(
+    <QueryClientProvider client={queryClient}>
+      <CancelConfirmDialog
+        counterpart="Far Settlement (Far Nation)"
+        onClose={onClose}
+        queryClient={queryClient}
+        route={makeRoute()}
+        traderCount={traderCount}
+      />
+    </QueryClientProvider>,
+  );
+  return { onClose, rpcMock };
+}
+
+describe("CancelConfirmDialog", () => {
+  beforeEach(() => {
+    requireSupabaseClient.mockReset();
+    toastError.mockReset();
+    toastSuccess.mockReset();
+  });
+
+  it("shows cancel dialog with counterpart name", () => {
+    renderDialog();
+    expect(
+      screen.getByRole("dialog", { name: "Cancel trade route?" }),
+    ).toBeDefined();
+    expect(screen.getByText(/Far Settlement \(Far Nation\)/)).toBeDefined();
+  });
+
+  it("happy path with no traders — shows plain success toast", async () => {
+    const user = userEvent.setup();
+    const { onClose } = renderDialog({ traderCount: 0 });
+
+    await user.click(screen.getByRole("button", { name: "Cancel route" }));
+
+    await waitFor(() => {
+      expect(toastSuccess).toHaveBeenCalledWith(
+        "Trade route cancelled.",
+        undefined,
+      );
+    });
+    expect(onClose).toHaveBeenCalled();
+  });
+
+  it("happy path with traders — includes trader count in toast", async () => {
+    const user = userEvent.setup();
+    renderDialog({ traderCount: 3 });
+
+    expect(screen.getByText(/3 traders/)).toBeDefined();
+
+    await user.click(screen.getByRole("button", { name: "Cancel route" }));
+
+    await waitFor(() => {
+      expect(toastSuccess).toHaveBeenCalledWith(
+        "Trade route cancelled. 3 traders were unassigned.",
+        undefined,
+      );
+    });
+  });
+
+  it("singular trader count uses 'trader was'", async () => {
+    const user = userEvent.setup();
+    renderDialog({ traderCount: 1 });
+
+    await user.click(screen.getByRole("button", { name: "Cancel route" }));
+
+    await waitFor(() => {
+      expect(toastSuccess).toHaveBeenCalledWith(
+        "Trade route cancelled. 1 trader was unassigned.",
+        undefined,
+      );
+    });
+  });
+
+  it("error branch — shows error toast on RPC failure", async () => {
+    const user = userEvent.setup();
+    const rpcMock = vi.fn((fn: string) => {
+      if (fn === "cancel_trade_route") {
+        return {
+          maybeSingle: vi.fn().mockResolvedValue({
+            data: null,
+            error: { message: "permission denied" },
+          }),
+        };
+      }
+      throw new Error(`Unexpected RPC: ${fn}`);
+    });
+    renderDialog({ rpcMock });
+
+    await user.click(screen.getByRole("button", { name: "Cancel route" }));
+
+    await waitFor(() => {
+      expect(toastError).toHaveBeenCalled();
+    });
+    expect(toastSuccess).not.toHaveBeenCalled();
+  });
+
+  it("keep button calls onClose without submitting", async () => {
+    const user = userEvent.setup();
+    const { onClose, rpcMock } = renderDialog();
+
+    await user.click(screen.getByRole("button", { name: "Keep" }));
+
+    expect(onClose).toHaveBeenCalled();
+    expect(rpcMock).not.toHaveBeenCalled();
+  });
+});
