+++ src/features/trade/components/SettlementTradeRoutesPanel/SettlementTradeRoutesPanel.test.tsx
@@ -3,7 +3,7 @@ import { render, screen, waitFor, within } from "@testing-library/react";
 import userEvent from "@testing-library/user-event";
 import { beforeEach, describe, expect, it, vi } from "vitest";
 
-import { SettlementTradeRoutesPanel } from "./SettlementTradeRoutesPanel";
+import { SettlementTradeRoutesPanel } from ".";
 
 const { requireSupabaseClient } = vi.hoisted(() => ({
   requireSupabaseClient: vi.fn<() => unknown>(),
